Fix visibility check not considering viewport
The visiblity check is not taking the viewport into account, which causes problems
with elements that are not "hidden", but are off-canvas.

Depending on the WebDriver service being used, this can result in you accidentally
interacting with elements that are not visible. Or, the WebDriver service may
throw an Exception when you try to interact with the element. In either case, we
typically want a reliable way to determine if the element is truly visible to the
user before we attempt to interact with it.

This implementation uses the getBoundingClientRect() method of the DOM Element to
ensure that the Element is completely within the page (i.e. not off-canvas).

Although you could argue that an Element that is only partially off-canvas is
still visible, certain WebDriver implementations will still not allow you to
interact withit.

I still need to implement tests for this feature. So a PR will not be opened until
that's complete.